7-1+1=7/src/main/java/domain/usecase/CreateWorkflowUseCase.java
package domain.usecase;

import domain.model.Workflow;

public class CreateWorkflowUseCase {
    private WorkflowRepository workflowRepository;

    public CreateWorkflowUseCase(WorkflowRepository workflowRepository) {
        this.workflowRepository = workflowRepository;
    }

    public void execute(CreateWorkflowInput input, CreateWorkflowOutput output) {
        Workflow workflow = new Workflow(input.getWorkflowName());
        workflowRepository.add(workflow);

        output.setWorkflowId(workflow.getWorkflowId());
    }
}
